lambda-minimal-working/requirements.txt
google-api-python-client==2.100.0
google-auth==2.41.1
google-auth-httplib2==0.1.0
